pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vllama"
version = "0.1.0"
description = "A vllm management tool with Ollama-like interface"
readme = {file = "README.md", content-type = "text/markdown; charset=UTF-8"}
requires-python = ">=3.11"
dependencies = [
    "typer>=0.20.0",
    "vllm>=0.11.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "pyyaml>=6.0",
    "httpx>=0.28.0",
    "psutil>=6.1.0",
    "nvidia-ml-py>=12.560.30",
    "rich>=13.9.0",
    "accelerate>=1.11.0",
    "huggingface-hub>=0.28.0",
    "flashinfer-python>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
]

[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
default = true

[project.scripts]
vllama = "vllama.cli:app"
